# docs/swagger/swagger.yaml
definitions:
  github_com_lusoris_venio_internal_models.CreateUserRequest:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: user@example.com
        maxLength: 255
        type: string
      first_name:
        example: John
        maxLength: 100
        type: string
      last_name:
        example: Doe
        maxLength: 100
        type: string
      password:
        example: SecurePass123!
        maxLength: 128
        minLength: 8
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  github_com_lusoris_venio_internal_models.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: SecurePass123!
        type: string
    required:
    - email
    - password
    type: object
  github_com_lusoris_venio_internal_models.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_lusoris_venio_internal_models.User'
    type: object
  github_com_lusoris_venio_internal_models.UpdateUserRequest:
    properties:
      avatar:
        type: string
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 100
        type: string
      is_active:
        type: boolean
      last_name:
        maxLength: 100
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  github_com_lusoris_venio_internal_models.User:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2026-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      last_name:
        example: Doe
        type: string
      updated_at:
        example: "2026-01-15T10:30:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  internal_api_handlers.ErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
  internal_api_handlers.HealthResponse:
    properties:
      services:
        additionalProperties:
          $ref: '#/definitions/internal_api_handlers.ServiceInfo'
        type: object
      status:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  internal_api_handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    required:
    - refresh_token
    type: object
  internal_api_handlers.RefreshTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  internal_api_handlers.ServiceInfo:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lusoris_venio_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lusoris_venio_internal_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lusoris_venio_internal_models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_lusoris_venio_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /api/v1/users:
    get:
      description: Get a paginated list of users
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_lusoris_venio_internal_models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: List users
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      description: Get user details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lusoris_venio_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_lusoris_venio_internal_models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lusoris_venio_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Update user
      tags:
      - users
  /health/live:
    get:
      description: Check if the application is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Liveness probe
      tags:
      - health
  /health/ready:
    get:
      description: Check if the application is ready to serve traffic
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_api_handlers.HealthResponse'
      summary: Readiness probe
      tags:
      - health
swagger: "2.0"